refacto: generic pdf-to-text + forceOcr
